Enum vgtk::VNode

pub enum VNode<Model: Component> {
    Object(VObject<Model>),
    Component(VComponent<Model>),
}

A node in the virtual component tree representing a Component or a Gtk widget.

Don't attempt to construct these directly. Use the gtk! macro instead.

Variants

Object(VObject<Model>)

Component(VComponent<Model>)

Methods

impl<Model: Component> VNode<Model>

pub fn empty() -> VNodeIterator<Model>

Make an empty iterator of VNodes.

Use this inside a code block in the gtk! macro to return an empty list of child nodes.