Trait vgtk::ext::ApplicationHelpers

pub trait ApplicationHelpers: GtkApplicationExt {
    fn new_unwrap(
        application_id: Option<&str>, 
        flags: ApplicationFlags
    ) -> Application { ... }

    fn get_child_accels<P: IsA<Action>>(&self, action: &P) -> Vec<GString> { ... }

    fn set_child_accels<P: IsA<Action>>(&self, action: &P, accels: &[&str]) { ... }
}

Helper trait for Application.

Provided methods

fn new_unwrap(
application_id: Option<&str>,
flags: ApplicationFlags
) -> Application

Construct a new Application and panic if it fails.

This is like Application::new, but returns an Application instead of an Option<Application>, so you can use it as a constructor in the gtk! macro.

fn get_child_accels<P: IsA<Action>>(&self, action: &P) -> Vec<GString>

fn set_child_accels<P: IsA<Action>>(&self, action: &P, accels: &[&str])

Loading content...