Function vgtk::message_dialog

pub async fn message_dialog<'_, W, S>(
    parent: Option<&'_ W>, 
    flags: DialogFlags, 
    message_type: MessageType, 
    buttons: ButtonsType, 
    is_markup: bool, 
    message: S
) -> ResponseType where
    W: IsA<Window>,
    S: AsRef<str>, 

Open a simple MessageDialog.

The arguments are passed directly to MessageDialog::new(). The is_markup flag, if set, will interpret the message as markup rather than plain text (see MessageDialog::set_markup()).

It returns a Future which will resolve to the ResponseType the user responds with.

Examples

vgtk::message_dialog(
    vgtk::current_window().as_ref(),
    DialogFlags::MODAL,
    MessageType::Error,
    ButtonsType::OkCancel,
    true,
    "<b>ERROR:</b> Unknown error."
).await;