Trait vgtk::Component

pub trait Component: Default + Unpin {
    type Message: Clone + Send + Debug + Unpin;
    type Properties: Clone + Default;
    fn view(&self) -> VNode<Self>;

    fn update(&mut self, _msg: Self::Message) -> UpdateAction<Self> { ... }

    fn create(_props: Self::Properties) -> Self { ... }

    fn change(&mut self, _props: Self::Properties) -> UpdateAction<Self> { ... }

    fn mounted(&mut self) { ... }

    fn unmounted(&mut self) { ... }
}

This is the trait your UI components should implement.

Associated Types

type Message: Clone + Send + Debug + Unpin

The type of messages you can send to the Component::update() function.

type Properties: Clone + Default

A struct type which holds the properties for your Component.

The gtk! macro will construct this from the attributes on the corresponding component element.

This is not relevant and should be set to () if you're writing a top level component.

Loading content...
Required methods

fn view(&self) -> VNode<Self>

Build a VNode tree to represent your UI.

This is called whenever the Component needs to re-render, and its UI state will be updated to reflect the VNode tree.

Loading content...
Provided methods

fn update(&mut self, _msg: Self::Message) -> UpdateAction<Self>

Process a Component::Message and update the state accordingly.

If you've made changes which should be reflected in the UI state, return UpdateAction::Render. This will call Component::view() and update the widget tree accordingly.

If you need to perform I/O, you can return UpdateAction::Defer, which will run an async action and call Component::update() again with its result.

Otherwise, return UpdateAction::None.

fn create(_props: Self::Properties) -> Self

Construct a new Component given a Component::Properties object.

This will never be called to construct a top level component. These will always be constructed using Default::default().

fn change(&mut self, _props: Self::Properties) -> UpdateAction<Self>

Update a Component's properties.

fn mounted(&mut self)

This method is called when the Component becomes visible to the user.

fn unmounted(&mut self)

This method is called when the Component becomes hidden or is removed entirely.

Loading content...